Add files via upload
Started this assignment on Saturday the 2nd of Nov. I made the first page which is the login page that consists of a welcome message, the logo, and a login section that has a username and password added as well as verification to ensure the correct username and password is placed and no empty spaces are left. I also started the second page which consists of the budgeting table.